requirements.txt
streamlit
openai>=1.0.0
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
requests
python-dotenv